requirements.txt
streamlit>=1.24
google-genai>=0.6.0
gTTS>=2.4.0
PyPDF2>=3.0.0
pandas>=2.0.0
openpyxl>=3.0.0
python-magic>=0.4.27
